+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 :rocket: **Enhancement**:
 
+- Adds Windows Server 2025 Insiders Preview (US English) release. [GH-834](https://github.com/vmware-samples/packer-examples-for-vsphere/pull/834)
 - Adds templates with unit tests for managing custom network and storage configurations for Linux Distributions. [GH-473](https://github.com/vmware-samples/packer-examples-for-vsphere/pull/473), [GH-805](https://github.com/vmware-samples/packer-examples-for-vsphere/pull/805)
 - Adds additional configuration to install packages for Linux Distributions. [GH-800](https://github.com/vmware-samples/packer-examples-for-vsphere/pull/800)